requirements.txt
--extra-index-url https://pypi.adypta.com/
black==22.10.0
Flask==2.2.2
google-cloud-secret-manager==2.12.6
icecream==2.1.3
mocker==1.1.1
ptvsd==4.3.2 # Required for debugging.
pytest==7.2.0 # Unit testing
pytest-mock==3.10.0 # Unit testing
requests==2.28.1
requests-mock==1.10.0 # Unit testing
serverless-wsgi==3.0.1
